Move bundler monkey patches
