+++ bundler/lib/dependabot/bundler/file_updater/lockfile_updater.rb
@@ -2,9 +2,9 @@
 
 require "bundler"
 
-require "bundler_definition_ruby_version_patch"
-require "bundler_definition_bundler_version_patch"
-require "bundler_git_source_patch"
+require "dependabot/monkey_patches/bundler/definition_ruby_version_patch"
+require "dependabot/monkey_patches/bundler/definition_bundler_version_patch"
+require "dependabot/monkey_patches/bundler/git_source_patch"
 
 require "dependabot/shared_helpers"
 require "dependabot/errors"
