+++ bundler/lib/dependabot/bundler/update_checker/shared_bundler_helpers.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require "bundler_definition_ruby_version_patch"
-require "bundler_definition_bundler_version_patch"
-require "bundler_git_source_patch"
+require "dependabot/monkey_patches/bundler/definition_ruby_version_patch"
+require "dependabot/monkey_patches/bundler/definition_bundler_version_patch"
+require "dependabot/monkey_patches/bundler/git_source_patch"
 
 require "excon"
