+++ bundler/lib/dependabot/monkey_patches/bundler/definition_bundler_version_patch.rb
