+++ bundler/lib/dependabot/monkey_patches/bundler/definition_ruby_version_patch.rb
