+++ bundler/lib/dependabot/monkey_patches/bundler/git_source_patch.rb
